third_party/blink/web_tests/platform/generic/fast/events/invalid-001-expected.txt

This test checks if checkValidity() correctly fires an invalid event on invalid form control elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

  
PASS Invalid control
PASS v[i].checkValidity() is false
PASS Invalid control
PASS v[i].checkValidity() is false
PASS Invalid control
PASS v[i].checkValidity() is false
PASS successfullyParsed is true

TEST COMPLETE
